+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## À venir
+
+### Corrections
+
+- Le tampon des documents PDF n'était plus appliqué que sur la première page (régression introduite en 5.1.17) #1573 #1587
+
 ## 5.1.17 - 2026-09-14
 
 ### Corrections