+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## À venir
+## 5.1.18 - 2026-09-17
 
-### Corrections
+### Correction
 
 - Le tampon des documents PDF n'était plus appliqué que sur la première page (régression introduite en 5.1.17) #1573 #1587